renamed project tree nav bar
