+++ frontend/src/layout/panel-layout/ProjectDropdownMenu.tsx
